fix: PostgreSQL-compatible SQL types in Flyway migrations; MODE=PostgreSQL in H2 test URLs
- DOUBLE → DOUBLE PRECISION in V1000, V1001, V1002
- Added MODE=PostgreSQL to H2 JDBC URL in test application.properties

Closes #66

+++ runtime/src/main/resources/db/migration/V1000__ledger_base_schema.sql
@@ -53,7 +53,7 @@ CREATE TABLE ledger_attestation (
     attestor_role    VARCHAR(100),
     verdict          VARCHAR(20)     NOT NULL,
     evidence         TEXT,
-    confidence       DOUBLE          NOT NULL,
+    confidence       DOUBLE PRECISION          NOT NULL,
     occurred_at      TIMESTAMP       NOT NULL,
     CONSTRAINT pk_ledger_attestation PRIMARY KEY (id),
     CONSTRAINT fk_attestation_entry FOREIGN KEY (ledger_entry_id) REFERENCES ledger_entry (id)
